v8.0.2

+++ bus-health/src/main/java/org/miaixz/bus/health/linux/driver/proc/ProcessStat.java
@@ -152,8 +152,8 @@ public final class ProcessStat {
      *
      * @return a map with socket as the key and pid as the value
      */
-    public static Map<Integer, Integer> querySocketToPidMap() {
-        Map<Integer, Integer> pidMap = new HashMap<>();
+    public static Map<Long, Integer> querySocketToPidMap() {
+        Map<Long, Integer> pidMap = new HashMap<>();
         for (File f : getPidFiles()) {
             int pid = Parsing.parseIntOrDefault(f.getName(), -1);
             File[] fds = getFileDescriptorFiles(pid);
@@ -162,7 +162,7 @@ public final class ProcessStat {
                 if (symLink != null) {
                     Matcher m = SOCKET.matcher(symLink);
                     if (m.matches()) {
-                        pidMap.put(Parsing.parseIntOrDefault(m.group(1), -1), pid);
+                        pidMap.put(Parsing.parseLongOrDefault(m.group(1), -1L), pid);
                     }
                 }
             }
